chore(client): move Techstack to icon
